Put front-end support for -recentEngines in XBoard
The engines are added to the menu. Problem is still that in XBoard
startup engines are given by command, rather than selected by combobox,
so the command rather than the nickname ends up in the menu. If this is
then selected, it cannot be found in the list. To limit the damage,
Fairy-Max is now pre-installed without nickname. -recentEngines is now
configured to 4.