Fix OK button of error popup
The error dialog popped up as a child of the board window, even when
DisplayError was called from a dialog. Now it checks if a transient
dialog is up, and if so, it pops up as a child of that, grabs focus,
and suppresses the system popdown of the transient dialog through a flag
dialogError until it is popped down itself.