Leave height of row that only contains buttons free
All elements of the GenericPopUp are still scaled to the text height,
including buttons appended behind them (browse or color buttons).
But we now make an exception for a row that only contains buttons,
as there is no harm in it being a little higher (and the buttons look
prettier that way).