Fix button grab for sweep selection
The piece-menu initialization limited the grab of button events to those
occurring inside the board window (presumably intended to limit them to
the piece menu, but that does not pop up in sweep-selection mode).
This is now suppressed when -pieceMenu is off, so that selection continues
outside the the board. (Which is very convenient for pieces pu close to the
edge.)