Undo translation of cps->which = first/second
The string in cps->which should not be translated, because it goes into
the debug file to label engine I/O, and many tools for processing
debug files (broadcasters, PGN extractors) rely on this being "first" or
"second". In all contexts where "first" and "second" need to be translated,
cps->which itself is already enclosed in _().