Keep XBoard windows attached on resize
Windows attached to bottom or right are displaced with the size increase.
Their perpendicular dimension is also scaled, to keep their edges at the same
fraction of the main window, unless they stick out. In the latter case the
fraction is clipped to 1, so that edge moves with the full size increase.