Fix seek-graph race condition
When a game started just after the user requested a seek graph, so that
the game start message would be received between sending 'sought' and
receiving the seek-ad list, an empty graph would pop up on the prompt,
and it was not possible to take it down, because the gameMode was no longer
IcsIdle.
  This patch cancels a sought command on receiving a board of your own game
to prevent this situation. In addition the system is made more robust by
always accepting a left-click as pop-down request, even when the seek graph
cannot be up, when it is up anyway.