Limit premove cancelling to click on from square
On second thought it might be better to limit cancellation of the
premove to up-clicks on the premove from-square. (Presumably this
will be a static click, but if the down-click was elsewhere, a new move
most have been entered, which would clear the existing premove anyway.)
This allows the user to already grab another piece that he wants to move
after premove, or when premove is rejected.