Fix focus of Game List
After focus was transferred to the filter text-edit, it was impossible
to get it back on the game list (to make the arrow keys working again).
A translation to left down-click is now used to assign focus, just as
it would for text edits. For unknown reason this only worked if the
translations were instated by Overriding, rather than Augmenting,
which again required the normal translation of downclick, Set(), which
selects the clicked item, to be explicitly mentioned as well.