Swap all engine-related options during engine loading
Some of the rare options were not swapped, and would thus not be
correctly loaded with the engine when -singleEngineList was true.
They are now also all reset to their default, before engine load,
to prevent options of one engine to linger on for a next that does
not specify them. The intentional exception is -firstHost.