Add logo widgets in main board window
New options -logoSize and -logoDir are added. The former enables logo display,
when non-zero and not too big. The latter is used by -autoLogo. The options
 -firstLogo and -secondLogo also work. Problem is that the clock height
is larger than the logo height, and for now a nasty trick in GenericPopUp
is used to position the next element (message field) w.r.t. the clocks,
rather than the last element of the row above (the logo): in case of a
SAME_ROW Graph element, 'last' is kept at the previous element.