Use in-place sweep-selection for click-click under-promotion
With detour under-promotion a piece cycles through all possible choices
during backwards dragging. But in a click-click move there is no dragging.
This used to be solved by letting the second click start a drag event of
the default promotion choice, which you could accept by immediately
releasing the mouse button, or drag around to change it first.
 The dragging in this case seemed a little unnatural, because the to-square
was already indicated by the down-click. So it has been changed now to the
promo choice appearing on the down-clicked to-square, followed by
EditPosition-like sweep-selection on that square when one wants to under-
promote. The move is always accepted no matter where the up-click takes
place, except when it takes place on the from-square (which offers a last-
resort abort possibiliy).
 Note that click-click promotion is essential in variants where it is not
obvious from the from-square alone that a promotion is imminent, but where
it depends on the to-square. (Such as Shogi.) There drag-drop detour under-
promotion is not possible.