Implement automatic partner observe
Starting a game, or starting to observe one, now automatically issues a
'pobserve' command if dual-board is switched on, so that the partner game
will pop up in the secondary board (assuming it is bughouse).
 The secondary board now displays the players in the title. (This (lightly)
breaks WinBoard again.) When the primary game ends, the secondary board
closes automatically. Initial sizing of the board is repaired, and now uses
the actual current board size, rather than the initial size of the main board.