Port chat boxes to XBoard
A new menu item Open Chat Window is added to the View menu. It opens a
window that can handle 5 chats. A row of buttons on top selects which is
the foreground chat partner, to which we type, and whose incoming messages
will be displayed in the memo. The leftmost 'button' is not really a button,
but a text edit, and the active partner is displayed there (so it can be
altered). Selecting another chat partner through a button, moves that partner
to the text-edit, and reshuffles the buttons. Incoming messages in a chat
that is not foreground make the corresponding button color orange.