Offer primitive paging in file browser
If there are more than 1000 files that would have to be displayed,
the number will be limited to 999, and a 'next page' entry will appear
at the bottom. Clicking that will give you the next group.
 This is an emergency measure only, as it interferes with sorting
(grouping is based on order inside the directory, i.e. kind of arbitrary,
and sorting is only done within a group.) There is no such thing for the
folder list yet. There is no way to go back to a previous page, other
than closing and re-opening the browser, or selecting another file type.