Keep track of virginity of back-rank pieces in variant seirawan
The forelast rank of the Board array is now used in VariantSChess to
contain flags that indicate virginity of the back-rank pieces, one bit
for white, the other for black. Legality checking of gatings now makes
use of this info. FENs mention all virgin pieces not implied by castling
rights in the castling field, by printing the corresponding file IDs,
as long as holdings are non-empty. The FEN reader now also understands
this format. EditPositionDone fakes virginity for all pieces that are
on their starting square.