Let windows stick to right display edge (WB)
Auxiliary windows docked to the board window from below or to the right, which are
pushed out of view when the board window resizes, will be trimmed to just touch the
diplay edge (if sufficient width or height remains). When the board window shrinks,
auxiliary windows touching the edge will be resized so they remain touching it.