Pay attention to extension of 'positional' arguments
An argument without preceding option name was assumed to be an -lgf argument.
This patch still does that by default, but if the argument ends in .fen, .epd,
.trn, .ini or .xop it is considered an -lpf, -tf or -opt argument.
  In addition, an implied -lpf or -lgf argument also imply the option -viewer,
which inserts the -viewerOptions string into the command line, and proesses it
as options. A new option -tourneyOptions is similarly inserted for an implied
-tf argument. These strings should contain all options needed to start XBoard
fully configured as game viewer or tourney playing agent on the mentioned file.