Add options -fe, -se, -is to load installed engines/ics from list
The volatile string options -fe, -se ('first/second engine') load the first
and second engine of the given nickname in the -firstChessProgramNames list,
including all options that were installed with it (similar to when in
WinBoard the that engine line had been selected through the Startup Dialog's
combobox). The option -is is similarly used to retrieve the line for an
ICS in the -icsNames list.