Allow Analyze Game to auto-step through entire game file
If the game index is set to auto-increment (-lgi -1), the Analyze Game
function will automatically load the next game of the current game file
(starting to auto-play/annotate that) when the end of the current game is
reached. This flushes the annotated game if auto-saving was set,
so that eventually all games in the file will be annotatd and saved.
  The game index is automatically set to auto-increment when we load
a file while in Analyze Game mode, but switching this mode on while a game
is already loaded will stop at the end of that game, as usual (if -lgi was
not explicitly set).