Allow ICS nickname as positional argument
If an argument is encountered that is not preceded by an option name,
it is now first checked if it corresponds to the nickname of an ICS
in -icsNames. If so, it is assumed to be the value of an -is option,
selecting -ics mode with that ICS and all the options installed with it.