Resolve conflict between -mps and -inc options
The age-old problem that a persistent valid -inc value will cause -mps
to be ignored is solved by the kludge of storing these two as a single
option (-mps) in the settings file, and declaring the other (-inc) volatile.
The range of negative numbers (invalid as -mps value) was available for
that, and now is used to store the negated value of -inc in msec.
For backward compatibility, specification of -inc causes interpretation
of -mps in the old way. The forcing of -inc to -1 in the master settings file
could be safely removed because of this change, so that the time increment
now is a persistent option.