Maintain in-memory copy of the board image
As reading back from the xlib-surface proves incredibly slow (cairo probably
converts the entire image from Xt format to cairo format when we set it
as a source, even though we only want to read a few dozen pixels), we now
keep an exact copy of the board image as a cairo bitmap in memory. This
restores the speed of the animation, but unfortunately leads to awful code.