Allow resizing of board window
In full png mode the size of the board window is no longer locked.
After resizing, the square size is calculated as the largest one that
would fit the window, and the window is then shrunk around it. To prevent
excessive redrawing during sizing, the redraw is delayed by putting it
in DragProc.