Make expose handler generic
The Graph Option now returns two handles in hitherto unused fields of its
Option struct: Option.choice is a cairo handle of a memory buffer the user
must draw to, while Option.textValue is the handle to the widget.
Expose events copy the memory buffer to the display for the exposed area.
A routine DrawExpose() is supplied to let the user generate expose events.