Increase drag delay too 200 msec
On my slow laptop it can take more than 100 msec to reallocate bitmaps
in the generic expose, with as a consequence that the DelayedDrag event
is not processed before the timeout expires. The timeout event then triggers
a redraw, which again takes so long that the timeout expires before the
event to reset the timer could be processed, and this just perpetuates
itself, leading to a large series of queued redraws, and a long waiting time
for an image to appear after the sizing stops.