Make xevalgraph.c backend
The only front-end stuff left in xevalgraph.c was an unnecessary reading
of the window size (which could be obtained from the WindowPlaceent struct),
and drawing (all cairo now). The drawing was moved to draw.c, and the
remaining part renamed nevalgraph.c.

An attempt to list them in a less chaotic way.