Append recent engines to engine menu
This is now entirely moved to the back-end, by doing it in the table
from which the menus are later generated. The special callback for engine
items is abandoned, and the euProc for such items is left NULL to flag
that RecentEngineEvent should be invoked with the applicable engine number.
To calculate the latter a global variable firstEngineItem is dynamically
set to the length of the regular engine menu.