Implement SAME_ROW layouting
A more general way for putting options on the same row was implemented:
The necessary hbox is always created before the first option to go into
it is processed (and discarded when a lone option is encountered). Options
that can be chained test for the existence of the box, and are packed
into it when it exists, in stead of filling the appropriate table cells
directly.