Make generic memo-event handler, and connect history callback
Large text widgets can now define a user handler in their Option.choice
field. If they do, agneric callback is added tothem, which catches mouse
events. Button release and pointer motion pass their coordinate and type
to the specified user routine. Button presses in addition retrieve
the text from the widget, set the cursor in it to the clicked point,
and convert it to a character position, also passed to the user.
  The user is responsible for ignoring events he is not interested in.