Fix main-window close
Hitting the window close button can cause a destroy or a delete event,
but catching the former apparently has as side effect that the dialog
is destroyed. Wich is wrong for the auxiliary windows, and lead to
errors if you try to pop them up again. The delete event does not have
such a side effect if the callback returns TRUE.
  The function of GenericPopUpCallback (handling OK and cancel button)
and GenericPopDown (handling the window close button) now have been
combined into one. In GTK both events conveniently are the same,
and pass the widget of the shell to be popped down.