Make board sizing work through subtracting fixed height
There is now no attempt to give a size to the board drawing area, so it
can later be sized as small as we want. The frame width is determined
by subtracting the nominal board size from the actual window size,
(because the clocks will have that width, and the board will startas 1x1),
while the vertical margin is determined as the difference between the
bactual window height, the drawing-area height minus a fudge of 25 pixels
to compesate for the erroneously allocated blank space below the board.