Re-instate X11 front-end
The front-end files had to be adapted for deletion of xedittags and
transfer of its routines to dialogs.c; the transfer of comment popups
to dialogs.c; transfer of drawing code from xevalgraph.c to draw.c.
It also had to provide wrappers for calls from callbacks it does not use
(the MemoProc) and duplicat case lables for the diversified Option types.