Allow external piece bitmaps and board border (WB)
The -pieceImageDirectory is now also used for WinBoard, to indicate
a directory of .bmp files, used to replace the built-ins. The outline
and solid bitmaps can be full-color. This allows winBoard to use the
WinBoard-zeta graphics. A new option -border can indicate a bitmap file
used to draw a half-square-wide border around the board.
  A white background has to be drawn underneath the black pieces, to avoid
color-interference with the board, just like happens for the white pieces.
These backgrounds now always use bright white, rather than -withePieceColor.

We still have to work on the scaling of the zeta (600x600) bitmaps to
the actual board size, as for some bitmaps the stratchblt is ugly.