Implement use of pause / resume protocol commands
Engines that implement feature pause=1 will be paused during their move
on a pause event; pondering engines will be controlled either through
'pause' or (if they don't support that) 'easy', which both should work
instantaneously. If the thinking engine does not support 'pause', the pausing
will be deferred to when it produces its move, and is implemented by not
relaying the move, and switch off pondering for both.