Prefer pause mode on pondering engine over 'easy'
If the thinking engine is stopped after its move because it does not
support pause, is is still better to use pause on the pondering engine
than to switch of its pondering with 'easy', as the latter might not
be instantaneous. Also switch off pondering in the thinking engine before
it gives the move, just in case it does not respond immediately to
easy / hard.