Implement aborting of games on engine request.
A new game-end command is added. For bakward compatibility it is disguised
as an old one, namely "1/2-1/2 {Engine Abort Request}". These resultDetails
are recognized as special, and rather than ending the game in a draw (which
GUIs not implementing this), the game will be completely 'expunged' from
the match or tourney it is part of, leaving no trace in the saveGameFile.