Allow use of ~ in pieceToChar for shadow pieces in any variant
The use of shadow-pieces for promoted Pawns was limited to Crazyhouse and
Bughouse. But this is not necessary, and in fact undesirable, as it made
it impossible to implement Crazyhouse960, which combines shadow promotions
with Fischer castling (only available in variant fischerandom)s, and thus
has to be played as 8x8+5_fischerandom, using the Crazyhouse -pieceToChaTable.