Catch unknown engine in tourney games
When an XBoard instance is playing on a tourney it can happen that it is
using a settings file with an engine list not yet containing a certain
participant, because that participant was just installed in the instance
that created the tourney file, but it did not save those modifications yet.
This error is now caught, and leads to an error popup and stopping of the
tourney play. This error was made non-fatal to allow the user to decide
if he wants to quit with or without saving the settings (which, after all,
does not seem up-to-date).