Draw both coords in a1
The lower-left board square needed two inscriptions, but DrawOneSquare()
only allowed one, so that the file ID would not be written. Now two strings
can be passed, and when align > 1 to indicate the alignment of the top
string, it is still checked if a bottom string is present as well, which
then is drawn with alignment 1 (= bottom right).
 The rank coord or piece count can now also be double digit.