Don't add PV moves on board clicking in AnalyzeMode
Playing PV moves by right-clicking the PV was only intended for clicks
in the engine Output window, selecting from multi-PV analysis output,
where by default it plays only the first move. PV walkig by right-clicking
the board, however, would start the walk at the end even in AnalyzeMode,
and thus add the entire PV. Which is probably not useful. So adding moves
in that case has now completely been suppressed, so it can be used for
harmless viewing of the latest PV.
 The behavior has also been made subject to a persistent Boolean option
-appendPV, configured to true, with a checkbox in General Options.
 A checkbox has also been added there for controlling the -variations option,
which is now also configured to true.