Better handling of undefined window parameters
Repositioning and resizing a newly created dialog was based on the width
parameter of the stored sizing info being valid. Now it is separately
tested if the position is valid, so that sizing and moving can be done
independently. This allows us to store position info, but not size info
for the slave board, so that it automatically will use the size of the
main board when first opened.