Fix sizing of slave board GTK
For lack of a known method to specify a default size for Graph widgets,
we have to use the same system as for the main board: create with 0x0 size,
and measure the size of the outer window to determine margins.
After every popup we then resize to the required board size plus these
margins. (Awful...)