fix bug #38401: xboard.texi doesn't build with texinfo-5.0 (tiny change)
xboard-4.7.0's documentation is not buildable with texinfo-5.0, it fails with a lot of error messages from makeinfo like
./xboard.texi:1026: @itemx must follow @item
makeinfo seems right, @itemx is not used as described in the texinfo manual.
Attached patch puts each item of the "Options" menu in it's own @subsection and formats the description of each item as a @table as neccessary. This way, xboard builds again and the manual doesn't look too bad.