Make skipping of unknown option smarter
To make XBoard better resistant to the presence of disbanded or future
options in its settings file, encountering those is no longer a fatal error.
But the error recovery only consisted of skipping the rest of the line.
This fails badly on string options with a multi-line value, which is
enclosed by {braces}. We now detect that case, to skip to } instead.