Fix saving of analyzed game
Anotated games after batch-mode Analyze Game were not auto-saved,
because the GameEnds() call supposed to do that was called with a NULL
argument for resultDetails. Now it is called with the resultDetails as
pre-existing in the gameInfo, or an empty string if none exist. This bug
was masked by the fact that in standard game ends (mate, etc.) the
analyzing engine would declare game end and supply a message.
 Also makes a distinction between Analyze Game and regular Analyze mode
when resuming analysis after loading a game; only the former now would
trigger a batch analysis.