Provide compatibility with Alien Edition setup command
The setup command of the Alien Edition has a board-size specifier
(possibly as prefix on a parent-variant name) between pieceToCharTable
and FEN. The regular version would choke on that, precluding engines
designed for the Alien protocol extension to have their setup processed
when running under regular XBoard. This patch makes it skip the extra info.