Fix GTK SetInsertPos
SetInsertPos() was not yet ported to GTK, with as a result that the cursor
was positioned at the start of a line recalled in the ICS Input Box,
rather than at the end. The code also called SetInsertPos to force scrolling
to that point in the text (which in Xaw is an automatic side effect). So for
multi-line text edits this function calls ScrollToCursor(). For some reason
this does not work when replacing texts in the Chat Box memo, when switching
chat, although it does work in the same memo when a line is added to the
existing text.