Fix Xaw file browser
The 'next page' entry of the Xaw file browser contained a \177 character
to keep at sorted at the bottom. This caused trouble for gettext.
Now the sorting range is adapted so the 'next page' is never part of it.
  More seriously, the listbox fill code had a wrong cast when testing for
an empty listbox on the first element, so that it only tested the low byte
of the pointer, in stead of the entire pointer for NULLness. As a result
the contents of the listboxes could suddenly disappear when even when it
was non-empty, because the first filename was allocated at an unlucky address.