Build in limited EPD capability for engine fingerprintig
The new option -finger FILENAME causes games to end after the first
move, where the initial position plus the move will then be written to
the specified file in EPD format. This can be used in match mode with
a set of start positions or opening lines to record the best moves of
an engine on a set of positions (as in CSVN fingerprinting).