Implement board-marker protocol
The commands 'lift SQ' and 'put SQ' are sent to the engine to indicate
the user selects or puts down pieces, so the engine can respond with
'highlight FEN' to put markers on the board. This implements an externally
driven -showTargetSquares feature. The engine has to enable this by
'feature highight=1' The highlight command is further ignored when
native showTargetSquares is on.
  A comma suffix on the square is used to indicate the user kept Ctrl
down during the to-square event reported by the put command.