Use highlight command to specify move legality
When a highlight command has been sent, LeftClick will only accept
to-square events when they occur on a square that highlight marked
as legal, i.e. indicated by a capital. The transparant 'color' T
is added to allow the engine to mark a to-square as legal without
putting any visible marker on it.