Allow engine to click squares on behalf of user
A new engine->GUI command 'click SQUARE' is added, which will trigger the
same events in the GUI as when the user clicked the mentioned SQUARE
(up+down). With legality testing off, this allows an engine to implement
one-click moving, in response to a 'lift' command issued by the GUI when
the user clicks the from-square.

A comma suffixed to the square is used to indicate the specified click
should be performed as if the user hat the Ctrl key pressed during it.